Add timestamp based resumption resolver
When moving backups to new topics (such as moving backups between
stage and prod) we often know the timestamp needed for resumption,
but offsets are tedious to establish.

This gives us the ability to get a resumption token from a timestamp,
so applications can offer a flag for resuming from a timestamp instead
of the committed resumption token.