Add message processor evictions
The message processor currently doesn't evict old messages. Originally
we had logic to discard blocks older than the reorg threshold, but
sometimes out-of-order delivery delivers messages more than the reorg
threshold out of order, which caused problems so we commented it out.

This relies on the OMP to notify the MP via `evictOlderThan(n)` to
delete messages that are beyond the reorg threshold of the OMP.