Prevent blocking on evictOlderThan
We were seeing thread lock up unexpectedly during block processing,
and pprof showed that the ProcessMessage() thread was getting stuck
in the LRU eviction, specifiically trying to send on evictOlderThan.

What was happening is that every time something got evicted from the
cache, it would get sent to evictCh. Once the cache is full, at least
one item gets evicted from the cache every iteration, but no more than
one item would be processed at the start of the Processmessage function.

Because of reorgs, sometimes multiple items are put on evictCh, but only
one item was being taken off, so eventually the channel filled up.

This fix addresses this in two ways:

First, it sends the item to the evictCh in a select/default statement,
so it can never block. In this use case, it's fine if a number gets
skipped, because we remove all pending batches less than the number on
the channel, and the next number will be larger, picking up the missing
item anyway.

Second, each time ProcessMessage() is executed it will exhaust the evictCh
before moving on to process the message, rather than just taking one
message.

Option 1 guarantees we won't see blocking, but that alone would mean the
queue eventually fills up and we have a lot of dangling batches. Option 2
eliminates the dangling batches, but in the event of a large enough reorg
it could still hypothetically fill up the channel and block, so option 1
is still needed.