Better out-of-order-delivery reorg handling
Our previous handling of out-of-order deliveries was an improvement,
but we still encountered an issue on Ropsten with a particularly
pernicious out-of-order scenario.

We reproduced this scenario with in some tests (refactoring the test
framework a bit to make it easier to set up different reorg scenarios)
and the solution ultimately proved to be queuing batches for future
assessment a bit more liberally than we had historically.