Refactor shutdownwg
The way the shutdownwg was structured, it could sometimes lead to
messages getting sent over a closed channel on shutdown if a message
was in the middle of processing when the shutdown initiated.

This change ensures that even messages that are in the middle of
processing complete before the shutdown waitgroup is released,
so no messages will be sent over a closed channel.