Better startup reorg handling
We've been seeing a behavior on startup where the blocks being added
lack parents in the target database, despite initializing with a
lastHash value.

It turns out that this was because if oldHead was nil, prepareReorg
would just emit the new block, and on startup it was likely that
oldHead would be loaded after a reorg at the same height.

This change avoids passing prepareReorg a nil block, and in that
scenario treats the block as though its weight was lower than
the pending block.

The net effect is that the immediate child of the initial lastHash
can be emitted before the lastHash block is identified, but the
initial lastHash block must be reorged out before any non-child
heavier blocks can be emitted.