More durable startup with empty hash
When starting with an empty hash, we don't care about history, we
just want to start from current. But if the first block we process
gets reorged out and we don't have the common ancestors in memory,
we end up not emitting any blocks at all.

This change detects that we started from an empty hash and will emit
blocks that we don't have parents for until we have 10 blocks in
memory. This should enable us to deal with most reorg-on-startup
scenarios.