Address Send / Unsubscribe / Close race condition
Previously I had assumed that any sub.Unsubscribe() followed by a
close(ch) for the channel corresponding to that subscription would
be safe, and that feed.Send() would never try to send over a channel
for which the corresponding Subscription had been closed.

However this assumption does not hold. The feed.Send() makes a copy
of the list of channels to send to, with no locks or other allowances
for the possibility that Unsubscribe() happens first, thus it is
possible to send on a closed channel.

I may revisit this to see if it's possible to make the feed/subscription
library more safe, but for now I'm making this use of it safe.