Unsubscribe after waitgroup is done
I believe the send on closed channel bug we were seeing was occurring
because we were unsubscribing a nullSubscription, subsequently replacing
sub with a feedSubscription that never got closed, then closing the
channel the feedSubscription was relying on.

Calling unsubscribe() after the initwg finishes ensures that we have
the version of `sub` that could be subscribed to a channel.