Don't block on websocket reconnects
The connection channel in websocket streams is intended to block
the Start() process until the connection is established, but sending
on the connection channel ends up causing reconnects to block.

This uses a select / case to skip over the send if the connected
channel blocks so that reconnects still happen.