Avoid unmarshalling batches in producers
We were unmarshalling batches to get at the block hash, but the hash
isn't in the marshalled data, it's in the kafka key, so unmarshalling
is unnecessary.

We were also using it to track the block number, but we weren't actually
doing anything with that block number. If we decided to do something
with the block number in the future we can add it back, but in the meantime
we can avoid unmarshalling and get the hash accurately.