Fix assumptions about batch id uniqueness
Batch IDs correspond to block numbers, and are only made unique
when combined with the block hash. The skipBatches cache wasn't
considering block hashes, only the block numbers. We should try
to broadcast a batch regardless of its skipBatches status, and
skipBatches should be used to tell us it's safe to ignore an
ErrUnknownBatch, not that we shouldn't broadcast the batch in
the first place.