Add PurgeReplayCache
Cardinal EVM has an rpc call `cardinal_reproduceBlocks` which is
intended to re-flush blocks to the stream if they were missing or
incorrect previously. But if a block in the specified range
appears in the cache of recent hashes, it won't get replayed when
we might need it to.

This addition allows the application to clear that cache prior to
replaying blocks.