Remove reorg logic from kafka transport
The reorg logic was well-intentioned but overly complicated. It never
worked properly and caused things to get hung up and sometimes crash.

This doesn't totally pull out the re-org logic: most of the producer
and consumer logic is still present, but the conditions that would
trigger that logic will never be able to occur. At some point we may
revisit making the logic work again, or may pull it out entirely.