Fix race condition around closing number channel
If we received two blocks at the same height, once the goroutine
for the first block to come in gets triggered on the hash channel
it closes the number channel. Most of the time, this would cause
the select/case for other goroutines at the same height to resolve
on the number channel closing and everything cleans up nicely, but
if the second block's hash channel comes in before the first block
closes the number channel, both goroutines could attempt to close
the number channel resulting in a panic.

This update adds a global number channel lock that is held just
long enough to check whether the number channel is still open and
then close it. The first goroutine to resolve on the hash channel
will close the channel while the second goroutine to resolve on
the hash channel waits for the lock. By the time the second one
holds the lock the number channel will be closed, so the internal
select case will resolve to the empty case and not try to reclose
the lock.