Don't double count empty entries
I don't know why I originally had Cardinal Streams check that the
previously stored value was non-empty before decrementing the prefix
counter. When we started getting empty values coming across, this
meant that those values got double counted, causing all sorts of problems.

I think it would have been perfectly okay to just check for the presence
of the value and not consider its length at all. But out of an abundance
of caution, and since I had already been doing a length check there that
might have existed for an important reason, I've made the length check
more sophisticated.

If the length of the stored value is 0 and the length of the message
value is non-zero, then we'll replace the stored value with the message
value but *not* decrement the prefix counter since this prefix has
already been accounted for.